// sportliner-lk-webapp/src/api/models/BranchOfficeItem.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Sportliner-lk-endpoint-api
 * Sportliner LK Endpoint API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';


/**
 * Branch office item
 * @export
 * @interface BranchOfficeItem
 */
export interface BranchOfficeItem {
    /**
     * 
     * @type {string}
     * @memberof BranchOfficeItem
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof BranchOfficeItem
     */
    address: string;
}

export function BranchOfficeItemFromJSON(json: any): BranchOfficeItem {
    return BranchOfficeItemFromJSONTyped(json, false);
}

export function BranchOfficeItemFromJSONTyped(json: any, ignoreDiscriminator: boolean): BranchOfficeItem {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': json['id'],
        'address': json['address'],
    };
}

export function BranchOfficeItemToJSON(value?: BranchOfficeItem | null): any {
    return BranchOfficeItemToJSONTyped(value, false);
}

export function BranchOfficeItemToJSONTyped(value?: BranchOfficeItem | null, ignoreDiscriminator: boolean = false): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
        'address': value.address,
    };
}