// sportliner-lk-webapp/src/api/models/Tariff.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Sportliner-lk-endpoint-api
 * Sportliner LK Endpoint API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const Tariff = {
    ONE_LESSON_PER_WEEK: 'ONE_LESSON_PER_WEEK',
    TWO_LESSONS_PER_WEEK: 'TWO_LESSONS_PER_WEEK',
    THREE_LESSONS_PER_WEEK: 'THREE_LESSONS_PER_WEEK',
    UNLIM: 'UNLIM'
} as const;
export type Tariff = typeof Tariff[keyof typeof Tariff];


export function TariffFromJSON(json: any): Tariff {
    return TariffFromJSONTyped(json, false);
}

export function TariffFromJSONTyped(json: any, ignoreDiscriminator: boolean): Tariff {
    return json as Tariff;
}

export function TariffToJSON(value?: Tariff | null): any {
    return value as any;
}